1-last_digit.c

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -14,7 +14,7 @@ int main(void)
 	srand(time(0));
 	n = rand() - RAND_MAX / 2;
 
-	int f;
+	float f;
 	f = n % 10;
 
 	if (f > 5)
